pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "clear-html"
description = "Clean and normalize HTML."
authors = [{name = "Zyte Group Ltd", email = "info@zyte.com"}]
readme = "README.rst"
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: Implementation :: CPython",
]
requires-python = ">=3.8"
dependencies = [
    "attrs>=20.3.0",
    "html-text>=0.5.2",
    "lxml>=4.4.3",
]
dynamic = ["version"]

[project.urls]
Source = "https://github.com/zytedata/clear-html"

[tool.setuptools.dynamic]
version = {attr = "clear_html.__version__"}

[tool.setuptools.package-data]
clear_html = ["py.typed"]

[tool.isort]
profile = "black"
multi_line_output = 3

[[tool.mypy.overrides]]
module = [
    "html_text.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "tests.*",
]

[tool.pytest.ini_options]
junit_family = "xunit2"
testpaths = [
    "clear_html",
    "tests",
]
addopts = "--doctest-modules"
